Changes for Linux (Ubuntu 14.04) build
