+++ osquery/CMakeLists.txt
@@ -29,11 +29,11 @@ SET_TARGET_PROPERTIES(osquery_static PROPERTIES OUTPUT_NAME osquery)
 ADD_EXECUTABLE(shell main/shell.cpp)
 TARGET_LINK_LIBRARIES(shell osquery_static)
 SET_TARGET_PROPERTIES(shell PROPERTIES OUTPUT_NAME osqueryi)
-SET_TARGET_PROPERTIES(shell PROPERTIES COMPILE_FLAGS "-std=c++11 -stdlib=libc++")
+SET_TARGET_PROPERTIES(shell PROPERTIES COMPILE_FLAGS ${OS_COMPILE_FLAGS})
 INSTALL(TARGETS shell DESTINATION bin)
 
 ADD_EXECUTABLE(daemon main/daemon.cpp)
 TARGET_LINK_LIBRARIES(daemon osquery_static)
 SET_TARGET_PROPERTIES(daemon PROPERTIES OUTPUT_NAME osqueryd)
-SET_TARGET_PROPERTIES(daemon PROPERTIES COMPILE_FLAGS "-std=c++11 -stdlib=libc++")
+SET_TARGET_PROPERTIES(daemon PROPERTIES COMPILE_FLAGS ${OS_COMPILE_FLAGS})
 INSTALL(TARGETS daemon DESTINATION bin)
