+++ osquery/core/CMakeLists.txt
@@ -1,10 +1,18 @@
+if(APPLE)
+  set (OS_CORE_SOURCE "darwin/test_util.cpp")
+else()
+  set (OS_CORE_SOURCE "")
+endif()
+
 ADD_LIBRARY(osquery_core
   init_osquery.cpp
   sqlite_util.cpp
   system.cpp
   test_util.cpp
+  ${OS_CORE_SOURCE}
   text.cpp
 )
+
 TARGET_LINK_LIBRARIES(osquery_core gflags)
 TARGET_LINK_LIBRARIES(osquery_core glog)
 TARGET_LINK_LIBRARIES(osquery_core gtest)
@@ -12,11 +20,11 @@ TARGET_LINK_LIBRARIES(osquery_core osquery_database)
 TARGET_LINK_LIBRARIES(osquery_core osquery_filesystem)
 TARGET_LINK_LIBRARIES(osquery_core osquery_sqlite)
 TARGET_LINK_LIBRARIES(osquery_core osquery_tables)
-SET_TARGET_PROPERTIES(osquery_core PROPERTIES COMPILE_FLAGS "-std=c++11 -stdlib=libc++")
+SET_TARGET_PROPERTIES(osquery_core PROPERTIES COMPILE_FLAGS ${OS_COMPILE_FLAGS})
 
 ADD_EXECUTABLE(status_tests status_tests.cpp)
 TARGET_LINK_LIBRARIES(status_tests gtest)
-SET_TARGET_PROPERTIES(status_tests PROPERTIES COMPILE_FLAGS "-std=c++11 -stdlib=libc++")
+SET_TARGET_PROPERTIES(status_tests PROPERTIES COMPILE_FLAGS ${OS_COMPILE_FLAGS})
 ADD_TEST(status_tests status_tests)
 
 ADD_EXECUTABLE(sqlite_util_tests sqlite_util_tests.cpp)
@@ -24,7 +32,7 @@ TARGET_LINK_LIBRARIES(sqlite_util_tests gtest)
 TARGET_LINK_LIBRARIES(sqlite_util_tests osquery_core)
 TARGET_LINK_LIBRARIES(sqlite_util_tests osquery_database)
 TARGET_LINK_LIBRARIES(sqlite_util_tests osquery_sqlite)
-SET_TARGET_PROPERTIES(sqlite_util_tests PROPERTIES COMPILE_FLAGS "-std=c++11 -stdlib=libc++")
+SET_TARGET_PROPERTIES(sqlite_util_tests PROPERTIES COMPILE_FLAGS ${OS_COMPILE_FLAGS})
 ADD_TEST(sqlite_util_tests sqlite_util_tests)
 
 ADD_EXECUTABLE(test_util_tests test_util_tests.cpp)
@@ -32,12 +40,12 @@ TARGET_LINK_LIBRARIES(test_util_tests gtest)
 TARGET_LINK_LIBRARIES(test_util_tests glog)
 TARGET_LINK_LIBRARIES(test_util_tests osquery_core)
 TARGET_LINK_LIBRARIES(test_util_tests osquery_sqlite)
-SET_TARGET_PROPERTIES(test_util_tests PROPERTIES COMPILE_FLAGS "-std=c++11 -stdlib=libc++")
+SET_TARGET_PROPERTIES(test_util_tests PROPERTIES COMPILE_FLAGS ${OS_COMPILE_FLAGS})
 ADD_TEST(test_util_tests test_util_tests)
 
 ADD_EXECUTABLE(text_tests text_tests.cpp)
 TARGET_LINK_LIBRARIES(text_tests gtest)
 TARGET_LINK_LIBRARIES(text_tests glog)
 TARGET_LINK_LIBRARIES(text_tests osquery_core)
-SET_TARGET_PROPERTIES(text_tests PROPERTIES COMPILE_FLAGS "-std=c++11 -stdlib=libc++")
+SET_TARGET_PROPERTIES(text_tests PROPERTIES COMPILE_FLAGS ${OS_COMPILE_FLAGS})
 ADD_TEST(text_tests text_tests)
