+++ osquery/core/darwin/PECocoaBackportsGlobal.h
