+++ osquery/core/test_util.h
@@ -14,6 +14,7 @@
 #include "osquery/config.h"
 #include "osquery/core.h"
 #include "osquery/database.h"
+#include "osquery/filesystem.h"
 
 namespace osquery {
 namespace core {
@@ -96,27 +97,6 @@ struct SplitStringTestData {
 
 // generate a set of test data to test osquery::core::splitString
 std::vector<SplitStringTestData> generateSplitStringTestData();
-
-// generate test content of a property list
-std::string getPlistContent();
-
-// generate test content of com.apple.alf
-std::string getALFcontent();
-
-// generate a test ptree of the content returned by getALFContent
-boost::property_tree::ptree getALFTree();
-
-// generate test content of an Info.plist file
-std::string getInfoPlistContent();
-
-// generate a test ptree of the content returned by getInfoPlistContent
-boost::property_tree::ptree getInfoPlistTree();
-
-// generate test content for a LaunchDaemon
-std::string getLaunchdContent();
-
-// generate a test ptree of the content returned by getLaunchdContent
-boost::property_tree::ptree getLaunchdTree();
 }
 }
