+++ osquery/devtools/CMakeLists.txt
@@ -3,4 +3,4 @@ TARGET_LINK_LIBRARIES(osquery_devtools glog)
 TARGET_LINK_LIBRARIES(osquery_devtools readline)
 TARGET_LINK_LIBRARIES(osquery_devtools osquery_sqlite)
 TARGET_LINK_LIBRARIES(osquery_devtools osquery_tables)
-SET_TARGET_PROPERTIES(osquery_devtools PROPERTIES COMPILE_FLAGS "-std=c++11 -stdlib=libc++")
+SET_TARGET_PROPERTIES(osquery_devtools PROPERTIES COMPILE_FLAGS ${OS_COMPILE_FLAGS})
