+++ osquery/filesystem/darwin/plist.mm
