+++ osquery/filesystem/darwin/plist_tests.cpp
@@ -5,7 +5,7 @@
 #include <glog/logging.h>
 #include <gtest/gtest.h>
 
-#include "osquery/core/test_util.h"
+#include "osquery/core/darwin/test_util.h"
 
 using namespace osquery::core;
 namespace pt = boost::property_tree;
