+++ osquery/tables/system/darwin/apps.cpp
