+++ osquery/tables/system/darwin/kextstat.cpp
