+++ osquery/tables/system/darwin/nvram.cpp
