+++ osquery/tables/system/linux/kernel_modules.cpp
@@ -1,0 +1,23 @@
+// Copyright 2004-present Facebook. All Rights Reserved.
+
+#include <boost/algorithm/string/trim.hpp>
+#include <boost/lexical_cast.hpp>
+
+#include <glog/logging.h>
+
+#include "osquery/core.h"
+#include "osquery/database.h"
+
+using namespace osquery::db;
+
+namespace osquery {
+namespace tables {
+
+QueryData genKernelModules() {
+	QueryData results;
+
+	return results;
+}
+
+}
+}
